Modification on login and details
