+++ src/main/java/wolox/training/config/SecurityConfig.java
@@ -7,11 +7,7 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import wolox.training.utils.CustomAuthenticationProvider;
-import wolox.training.utils.ErrorConstants;
-import wolox.training.utils.VariablesConstants;
+import wolox.training.utils.AuthProviderUser;
 
 @Configuration
 @EnableWebSecurity
@@ -19,7 +15,7 @@ public class SecurityConfig extends
     WebSecurityConfigurerAdapter {
 
   @Autowired
-  private CustomAuthenticationProvider authProvider;
+  private AuthProviderUser authProvider;
 
   @Override
   protected void configure(final AuthenticationManagerBuilder auth) {
