+++ src/main/java/wolox/training/controllers/ApiController.java
@@ -1,18 +1,8 @@
 package wolox.training.controllers;
 
 import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import io.swagger.models.Model;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import wolox.training.errors.book.BookHttpErrors;
 import wolox.training.errors.user.UserHttpErrors;
@@ -21,8 +11,6 @@ import wolox.training.models.Book;
 import wolox.training.models.User;
 import wolox.training.repositories.BookRepository;
 import wolox.training.repositories.UserRepository;
-import wolox.training.utils.AuthProviderUser;
-import wolox.training.utils.CustomAuthenticationProvider;
 
 public abstract class ApiController {
