+++ src/main/java/wolox/training/utils/AuthProviderUser.java
@@ -1,6 +1,16 @@
 package wolox.training.utils;
 
-public class AuthProviderUser {
+import java.util.ArrayList;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthProviderUser implements
+    AuthenticationProvider {
+
   private String username;
   private String password;
 
@@ -25,5 +35,21 @@ public class AuthProviderUser {
 
   public String getPassword(){
     return password;
+  }
+
+  @Override
+  public Authentication authenticate(Authentication authentication)
+      throws AuthenticationException {
+
+    String name = authentication.getName();
+    String password = authentication.getCredentials().toString();
+
+    return new UsernamePasswordAuthenticationToken(
+          name, password, new ArrayList<>());
+  }
+
+  @Override
+  public boolean supports(Class<?> authentication) {
+    return authentication.equals(UsernamePasswordAuthenticationToken.class);
   }
 }
